proto: full schema parity with soylib (gift, chat, member, user, badge, public_area)
split proto module into types/user/gift_types/messages/messages_ext. expand User
to ~75 fields (is_subscribe, is_follower, sec_uid, pay_grade, fans_club_info,
user_attr, ...). fix tag mappings against captured wire data (Like, Caption,
Member, Social, RoomUserSeq, Contributor types corrected). add CombineBadge
with BadgeText (distinct from TextBadge), PublicAreaMessageCommon with
PortraitInfo metrics, full Gift sub-types (GiftIMPriority, GiftMonitorInfo,
SponsorshipInfo, MatchInfo, TextEffect, InteractiveGiftInfo, etc).

verified 100% decode rate on live capture replays.

+++ src/helpers/like_accumulator.rs
@@ -24,13 +24,13 @@ use crate::structs::proto::messages::WebcastLikeMessage;
 #[derive(Clone, Debug)]
 pub struct LikeStats {
     pub event_like_count: i32,
-    pub total_like_count: i32,
+    pub total_like_count: i64,
     pub accumulated_count: i64,
     pub went_backwards: bool,
 }
 
 pub struct LikeAccumulator {
-    max_total: i32,
+    max_total: i64,
     accumulated: i64,
 }
